requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
fonttools==4.59.2
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
# Emergency Reporting System Dependencies
twilio==8.10.0
phonenumbers==8.13.0
geopy==2.4.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
